requirements.txt
aabsl-py==0.9.0
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
cymem==2.0.3
future==0.18.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-cloud-storage==1.29.0
grpcio==1.30.0
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
murmurhash==1.0.2
numpy==1.18.4
oauthlib==3.1.0
plac==1.1.3
preshed==3.0.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytorch-lightning==0.7.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
spacy==2.3.0
srsly==1.0.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
thinc==7.4.1
torch==1.5.1
tqdm==4.47.0
urllib3==1.25.9
wasabi==0.7.0
Werkzeug==1.0.1
zipp==3.1.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.0/en_core_web_sm-2.3.0.tar.gz